Separate javascript sdk from example
