+++ .gitignore
@@ -7,3 +7,5 @@ protobuf-3.6.0
 __pycache__
 
 *.bin
+*/junk/*
+node_modules
