+++ javascript/example/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "sensr_social_distancing_example",
+  "version": "1.0.0",
+  "description": "",
+  "main": "social_distancing.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "mathjs": "^7.0.0"
+  }
+}
