+++ javascript/javascript_sdk/gen_proto.sh
@@ -2,7 +2,7 @@
 
 SRC_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 OUT_DIR="${SRC_DIR}/js_proto"
-PROTO_DIR="${SRC_DIR}/../proto"
+PROTO_DIR="${SRC_DIR}/../../proto"
 
 if [ ! -d $OUT_DIR ]; then
   mkdir $OUT_DIR
