+++ javascript/javascript_sdk/troubleshooting.md
